# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface UpdateProfileBody {
  displayName?: string;
  signatureImage?: string;
  email?: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email?: string;
  role: string;
  displayName: string;
  signatureImage?: string;
}

export interface UserRecord {
  id: number;
  username: string;
  email?: string;
  role: string;
  displayName: string;
  createdAt: string;
}

export type CreateUserBodyRole = typeof CreateUserBodyRole[keyof typeof CreateUserBodyRole];


export const CreateUserBodyRole = {
  user: 'user',
  admin: 'admin',
  super_admin: 'super_admin',
} as const;

export interface CreateUserBody {
  username: string;
  email?: string;
  password: string;
  role?: CreateUserBodyRole;
  displayName: string;
}

export type UpdateUserBodyRole = typeof UpdateUserBodyRole[keyof typeof UpdateUserBodyRole];


export const UpdateUserBodyRole = {
  user: 'user',
  admin: 'admin',
  super_admin: 'super_admin',
} as const;

export interface UpdateUserBody {
  role?: UpdateUserBodyRole;
  displayName?: string;
  email?: string;
}

export interface RejectBody {
  reason: string;
}

export interface LineItem {
  id: string;
  description: string;
  quantity: number;
  unitPrice: number;
}

export type QuotationPayloadDocumentType = typeof QuotationPayloadDocumentType[keyof typeof QuotationPayloadDocumentType];


export const QuotationPayloadDocumentType = {
  quotation: 'quotation',
  invoice: 'invoice',
  receipt: 'receipt',
  delivery_note: 'delivery_note',
  sale_order: 'sale_order',
} as const;

export type QuotationPayloadDiscountType = typeof QuotationPayloadDiscountType[keyof typeof QuotationPayloadDiscountType];


export const QuotationPayloadDiscountType = {
  fixed: 'fixed',
  percentage: 'percentage',
} as const;

export interface QuotationPayload {
  documentType?: QuotationPayloadDocumentType;
  quotationNumber: string;
  date: string;
  validUntil: string;
  clientName: string;
  companyName?: string;
  address?: string;
  email?: string;
  phone?: string;
  items: LineItem[];
  discountType: QuotationPayloadDiscountType;
  discountValue: number;
  applyTax: boolean;
  notes?: string;
  preparedBy?: string;
  signatureImage?: string;
}

export type QuotationRecord = QuotationPayload & {
  id: number;
  status: string;
  submittedById?: number;
  approvedById?: number;
  approvedAt?: string;
  rejectionReason?: string;
  createdAt: string;
  updatedAt: string;
};

export interface QuotationSummary {
  id: number;
  documentType?: string;
  quotationNumber: string;
  clientName: string;
  companyName?: string;
  date: string;
  validUntil?: string;
  status: string;
  preparedBy?: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export interface InventoryItemPayload {
  name: string;
  description?: string;
  category?: string;
  unit?: string;
  unitPrice: number;
  isActive?: boolean;
}

export type InventoryItemRecord = InventoryItemPayload & {
  id: number;
  createdAt: string;
  updatedAt: string;
};

export type LeadPayloadStatus = typeof LeadPayloadStatus[keyof typeof LeadPayloadStatus];


export const LeadPayloadStatus = {
  new: 'new',
  contacted: 'contacted',
  qualified: 'qualified',
  lost: 'lost',
} as const;

export interface LeadPayload {
  clientName: string;
  companyName?: string;
  email?: string;
  phone?: string;
  productInterest?: string;
  status?: LeadPayloadStatus;
  notes?: string;
}

export type LeadRecord = LeadPayload & {
  id: number;
  createdAt: string;
  updatedAt: string;
};

export type FollowUpPayloadStatus = typeof FollowUpPayloadStatus[keyof typeof FollowUpPayloadStatus];


export const FollowUpPayloadStatus = {
  pending: 'pending',
  done: 'done',
  cancelled: 'cancelled',
} as const;

export interface FollowUpPayload {
  leadId: number;
  scheduledDate: string;
  notes?: string;
  status?: FollowUpPayloadStatus;
}

export type FollowUpRecord = FollowUpPayload & {
  id: number;
  createdAt: string;
};

export type Logout200 = {
  ok: boolean;
};

export type ChangePassword200 = {
  ok: boolean;
};

export type ListInventoryItemsParams = {
activeOnly?: boolean;
};