# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/profile:
    put:
      operationId: updateProfile
      tags: [auth]
      summary: Update current user display name and signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    put:
      operationId: changePassword
      tags: [auth]
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /quotations:
    get:
      operationId: listQuotations
      tags: [quotations]
      summary: List all saved quotations
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuotationSummary"
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Save a new quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationPayload"
      responses:
        "201":
          description: Created quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotations/{id}:
    get:
      operationId: getQuotation
      tags: [quotations]
      summary: Get a single quotation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateQuotation
      tags: [quotations]
      summary: Update an existing quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationPayload"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuotation
      tags: [quotations]
      summary: Delete a quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotations/{id}/submit:
    post:
      operationId: submitQuotation
      tags: [quotations]
      summary: Submit a quotation for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotations/{id}/approve:
    post:
      operationId: approveQuotation
      tags: [quotations]
      summary: Approve a quotation (admin/super_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotations/{id}/reject:
    post:
      operationId: rejectQuotation
      tags: [quotations]
      summary: Reject a quotation (admin/super_admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBody"
      responses:
        "200":
          description: Quotation rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuotationRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    UpdateProfileBody:
      type: object
      properties:
        displayName:
          type: string
        signatureImage:
          type: string
        email:
          type: string
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
      required: [currentPassword, newPassword]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        displayName:
          type: string
        signatureImage:
          type: string
      required: [id, username, role, displayName]
...
    RejectBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]
    LineItem:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
      required: [id, description, quantity, unitPrice]
    QuotationPayload:
      type: object
      properties:
        documentType:
          type: string
          enum: [quotation, invoice, receipt, delivery_note, sale_order]
        quotationNumber:
          type: string
        date:
          type: string
        validUntil:
          type: string
        clientName:
          type: string
        companyName:
          type: string
        address:
          type: string
        email:
          type: string
        phone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LineItem"
        discountType:
          type: string
          enum: [fixed, percentage]
        discountValue:
          type: number
        applyTax:
          type: boolean
        notes:
          type: string
        preparedBy:
          type: string
        signatureImage:
          type: string
      required:
        - quotationNumber
        - date
        - validUntil
        - clientName
        - items
        - discountType
        - discountValue
        - applyTax
    QuotationRecord:
      allOf:
        - $ref: "#/components/schemas/QuotationPayload"
        - type: object
          properties:
            id:
              type: integer
            status:
              type: string
            submittedById:
              type: integer
            approvedById:
              type: integer
            approvedAt:
              type: string
            rejectionReason:
              type: string
            createdAt:
              type: string
            updatedAt:
              type: string
          required: [id, status, createdAt, updatedAt]
    QuotationSummary:
      type: object
      properties:
        id:
          type: integer
        documentType:
          type: string
        quotationNumber:
          type: string
        clientName:
          type: string
        companyName:
          type: string
        date:
          type: string
        validUntil:
          type: string
        status:
          type: string
        preparedBy:
          type: string
        createdAt:
          type: string
      required: [id, quotationNumber, clientName, date, createdAt, status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...